refactor: separate cloud and self-hosted components
