+++ src/components/pages/pricing/hero-with-cards/cloud/index.js
@@ -1,0 +1,1 @@
+export { default } from './cloud';
