+++ src/components/pages/pricing/hero-with-cards/self-hosted/index.js
@@ -1,0 +1,1 @@
+export { default } from './self-hosted';
